encapsulate localStorage access within RootServerApi
