+++ src/resources/js/routes/ProtectedLoginRoute.tsx
@@ -1,5 +1,4 @@
 import { Navigate } from 'react-router-dom';
-import { loadToken } from '../helpers/utils';
 import RootServerApi from '../RootServerApi';
 
 type Props = {
@@ -7,8 +6,6 @@ type Props = {
 };
 
 export const ProtectedLoginRoute = ({ children }: Props) => {
-  loadToken();
-
   if (RootServerApi.isLoggedIn) {
     return <Navigate to='/' />;
   }
